@@ -21,10 +21,17 @@
     "discord-api-types": "^0.37.42",
     "discord.js": "^14.0.4-dev.1658491502-74ec7be.0",
     "libsodium-wrappers": "^0.7.10",
-    "ts-node": "^10.9.1",
-    "typescript": "^5.0.4"
+    "ts-node": "^10.9.1"
   },
   "devDependencies": {
-    "prettier": "^2.5.1"
+    "@typescript-eslint/eslint-plugin": "^6.7.2",
+    "@typescript-eslint/parser": "^6.7.2",
+    "eslint": "^8.50.0",
+    "eslint-config-standard-with-typescript": "^39.0.0",
+    "eslint-plugin-import": "^2.28.1",
+    "eslint-plugin-n": "^16.1.0",
+    "eslint-plugin-promise": "^6.1.1",
+    "prettier": "^2.5.1",
+    "typescript": "^5.2.2"
   }
 }
