@@ -1,60 +1,47 @@
-const { FACEIT_API_KEY } = require("../../config.json");
-import axios, { AxiosRequestConfig } from "axios";
+import { FACEIT_API_KEY } from "../../config.json";
+import axios from "axios";
 import { FaceitPlayer } from "./types";
 
-/**
- * Authorization headers
- */
-const options: AxiosRequestConfig<any> = {
+const options = {
   headers: {
     "Content-Type": "application/json",
     Authorization: `Bearer ${FACEIT_API_KEY}`,
   },
 };
-/**
- * Gets FACEIT player data from username
- * @returns player data
- */
-const searchPlayer = async (username: string): Promise<FaceitPlayer> => {
+
+export const searchPlayer = async (username: string): Promise<FaceitPlayer> => {
   const query = `https://open.faceit.com/data/v4/players?nickname=${username}&game=csgo`;
   return axios.get(query, options);
 };
-/**
- * Gets FACEIT player data from steamID
- * @returns player data
- */
-const searchPlayerFromSteamID = async (
+
+export const searchPlayerFromSteamID = async (
   steamID: string
 ): Promise<FaceitPlayer> => {
   const query = `https://open.faceit.com/data/v4/players?game=csgo&game_player_id=${steamID}`;
   return axios.get(query, options);
 };
-/**
- * Gets FACEIT player all time stats
- */
-const searchPlayerStats = async (username: string) => {
+
+export const searchPlayerStats = async (username: string) => {
   const player_id = (await searchPlayer(username))?.data?.player_id;
   const query = `https://open.faceit.com/data/v4/players/${player_id}/stats/csgo`;
   return axios.get(query, options);
 };
-/**
- * Gets FACEIT match history from faceit username
- * @returns array of matches
- */
-const getMatchHistory = async (username: string) => {
+
+export const getMatchHistory = async (username: string) => {
   const playerId = (await searchPlayer(username))?.data?.player_id;
   const query = `https://api.faceit.com/stats/api/v1/stats/time/users/${playerId}/games/csgo?size=100`;
   return await paginatedRequest(query, playerId, 1);
 };
 
-const paginatedRequest = async (
+export const paginatedRequest = async (
   query: string,
   playerId: string,
   page: number,
   data: number[] = []
 ) => {
   await axios.get(query).then((res) => {
-    data.push.apply(data, res?.data);
+    // eslint-disable-next-line prefer-spread
+    data.push(...res.data);
 
     if (res?.data && res.data?.length != 0) {
       return paginatedRequest(
@@ -66,11 +53,4 @@ const paginatedRequest = async (
     }
   });
   return data;
-};
-
-module.exports = {
-  searchPlayer,
-  searchPlayerFromSteamID,
-  searchPlayerStats,
-  getMatchHistory,
 };
